requirements.txt
ftfy
regex
tqdm
torch
torchvision
pillow